-- usinfy method save for persist entities - fix test after Adding ResType

+++ src/main/java/org/shved/webacs/dao/IResTypeDAO.java
@@ -18,7 +18,7 @@ public interface IResTypeDAO {
 
     void delete(ResType resType);
 
-    void deleteById(Long id);
+    void deleteById(Integer id);
 
     ResType findByName(String name);
